jarvis_assistant/requirements.txt
# Core dependencies
python-dotenv==1.0.0

# LLM APIs
anthropic==0.39.0
openai==1.54.0

# Vector database and embeddings
chromadb==0.5.11
sentence-transformers==3.3.1

# Utilities
tqdm==4.66.1
